moved login functionality to its own file
