+++ Server/f1/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from .views import user_view
+from .views import user_view, login_view
 from drf_spectacular.views import SpectacularAPIView, SpectacularRedocView, SpectacularSwaggerView
 
 urlpatterns = [
@@ -9,9 +9,9 @@ urlpatterns = [
     path('api/schema/redoc/', SpectacularRedocView.as_view(url_name='schema'), name='redoc'),
     # APIS
     path('user/<int:user_id>', user_view.UserView.as_view()),
-    path('register/user/', user_view.RegisterUserView.as_view()),
-    path('login/user/', user_view.LoginUserSerializer.as_view()),
-    path('logout/user/', user_view.LoginUserSerializer.as_view()),
+    path('register/user/', login_view.RegisterUserView.as_view()),
+    path('login/user/', login_view.LoginUserSerializer.as_view()),
+    path('logout/user/', login_view.LogoutUserView.as_view()),
     path('user/<int:user_id>/leagues/', user_view.UserLeagues.as_view()),
     path('user/<int:user_id>/league/', user_view.UserCreateLeague.as_view()),
     path('user/<int:user_id>/league/<int:league_id>', user_view.UserLeague.as_view())
